changed sample_collision to be a particle method; inlined rng sampling
